agent/pyproject.toml
[project]
name = "tota-agent"
version = "0.1.0"
requires-python = ">=3.10, <3.15"
dependencies = [
    "livekit-agents[sarvam,google]~=1.3",
    "python-dotenv",
]
